Move ExponentialBackoff to :time
