+++ platforms/core-runtime/base-services/build.gradle.kts
@@ -52,11 +52,12 @@ errorprone {
 dependencies {
     api(projects.concurrent)
     api(projects.javaLanguageExtensions)
-    api(projects.time)
     api(project(":hashing"))
     api(project(":build-operations"))
     api(libs.jsr305)
     api(libs.guava)
+
+    implementation(projects.time)
 
     implementation(libs.asm)
     implementation(libs.commonsIo)
