+++ platforms/software/security/build.gradle.kts
@@ -18,6 +18,7 @@ dependencies {
     api(libs.jsr305)
 
     implementation(projects.javaLanguageExtensions)
+    implementation(projects.time)
     implementation(project(":base-services"))
     implementation(project(":functional"))
     implementation(project(":logging-api"))
